requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
scikit-learn
tqdm
